Module escrow::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Type Definitions

Blocks

Storage type is [StorageMap] with key type T :: AccountId and value type bool.

Epoch

Storage type is [StorageValue] with value type T :: Nonce.

Limits

Storage type is [StorageMap] with key type T :: AccountId and value type (BlockNumberFor < T >, BlockNumberFor < T >).

Locked

Storage type is [StorageMap] with key type T :: AccountId and value type LockedBalance < BalanceOf < T >, BlockNumberFor < T > >.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PointHistory

Storage type is [StorageMap] with key type T :: Nonce and value type Point < BalanceOf < T >, BlockNumberFor < T > >.

Reserved

Storage type is [StorageMap] with key type T :: AccountId and value type BalanceOf < T >.

SlopeChanges

Storage type is [StorageMap] with key type BlockNumberFor < T > and value type BalanceOf < T >.

UserPointEpoch

Storage type is [StorageMap] with key type T :: AccountId and value type T :: Nonce.

UserPointHistory

Storage type is [StorageDoubleMap] with key1 type T :: AccountId, key2 type T :: Nonce and value type Point < BalanceOf < T >, BlockNumberFor < T > >.